Types in alligator.typing
