+++ alligator/processors.py
@@ -1,42 +1,15 @@
 import hashlib
 import traceback
 from collections import defaultdict
-from dataclasses import dataclass
-from typing import Any, Dict, List, Optional
+from typing import Dict, List
 
 import pandas as pd
 
 from alligator.feature import Feature
 from alligator.fetchers import CandidateFetcher, LiteralFetcher, ObjectFetcher
 from alligator.mongo import MongoWrapper
+from alligator.typing import Entity, RowData
 from alligator.utils import tokenize_text
-
-
-@dataclass
-class Entity:
-    """Represents a named entity from a table cell."""
-
-    value: str
-    row_index: Optional[int]
-    col_index: str  # Stored as string for consistency
-    context_text: str
-    correct_qid: Optional[str] = None
-    fuzzy: bool = False
-
-
-@dataclass
-class RowData:
-    """Represents a row's data with all necessary context."""
-
-    doc_id: str
-    row: List[Any]
-    ne_columns: Dict[str, str]
-    lit_columns: Dict[str, str]
-    context_columns: List[str]
-    correct_qids: Dict[str, str]
-    row_index: Optional[int]
-    context_text: str
-    row_hash: str
 
 
 class RowBatchProcessor:
